requirements.txt
langchain
langchain-core
langchain-classic
langchain-community
langchain-mistralai
langchain-huggingface
langchain-google-genai
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
chainlit
